+++ src/events/configUpdateEntry.js
@@ -3,12 +3,12 @@ const gateways = ['users', 'clientStorage'];
 
 module.exports = class extends Event {
 
-	run(configs) {
-		if (this.client.shard && gateways.includes(configs.gateway.type)) {
+	run(settings) {
+		if (this.client.shard && gateways.includes(settings.gateway.type)) {
 			this.client.shard.broadcastEval(`
 				if (this.shard.id !== ${this.client.shard.id}) {
-					const entry = this.gateways.${configs.gateway.type}.cache.get('${configs.id}');
-					if (entry) entry._patch(${JSON.stringify(configs)});
+					const entry = this.gateways.${settings.gateway.type}.cache.get('${settings.id}');
+					if (entry) entry._patch(${JSON.stringify(settings)});
 				}
 			`);
 		}