+++ src/lib/structures/SQLProvider.js
@@ -61,7 +61,7 @@ class SQLProvider extends Provider {
 	/**
 	 * Parse the gateway input for easier operation
 	 * @since 0.5.0
-	 * @param {(ConfigurationUpdateResultEntry[]|Array<Array<string>>|Object<string, *>)} [updated] The updated entries
+	 * @param {(SettingsUpdateResultEntry[]|Array<Array<string>>|Object<string, *>)} [updated] The updated entries
 	 * @param {boolean} [resolve=true] Whether this should resolve the values using QueryBuider#resolve or not
 	 * @returns {Array<any[]>}
 	 * @protected
@@ -79,11 +79,11 @@ class SQLProvider extends Provider {
 			else if (first.data && first.piece) this._parseGatewayInput(updated, keys, values, resolve);
 
 			// Unknown overload, throw
-			else throw new TypeError(`Expected void, [k, v][], ConfigurationUpdateResult[], or an object literal. Got: ${new Type(updated)}`);
+			else throw new TypeError(`Expected void, [k, v][], SettingsUpdateResult[], or an object literal. Got: ${new Type(updated)}`);
 
 			return [keys, values];
 		}
-		if (!isObject(updated)) throw new TypeError(`Expected void, [k, v][], ConfigurationUpdateResult[], or an object literal. Got: ${new Type(updated)}`);
+		if (!isObject(updated)) throw new TypeError(`Expected void, [k, v][], SettingsUpdateResult[], or an object literal. Got: ${new Type(updated)}`);
 
 		return objectToTuples(updated);
 	}
@@ -151,8 +151,8 @@ class SQLProvider extends Provider {
 	}
 
 	/**
-	 * Parse the ConfigurationUpdateResultEntry[] overload
-	 * @param {ConfigurationUpdateResultEntry[]} updated The updated keys
+	 * Parse the SettingsUpdateResultEntry[] overload
+	 * @param {SettingsUpdateResultEntry[]} updated The updated keys
 	 * @param {string[]} keys The keys to update
 	 * @param {any[]} values The values to update
 	 * @param {boolean} [resolve = true] Whether this should resolve the values using QueryBuider#resolve or not